+++ sdk/devcenter/azure-mgmt-devcenter/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Release History
 
+## 1.0.0b3 (2022-10-27)
+
+### Features Added
+
+  - Model Catalog has a new parameter sync_state
+  - Model CatalogProperties has a new parameter sync_state
+  - Model OperationStatus has a new parameter operations
+  - Model OperationStatus has a new parameter resource_id
+
+### Breaking Changes
+
+  - Client name is changed from `DevCenterClient` to `DevCenter`
+  - Parameter status of model OperationStatus is now required
+
 ## 1.0.0b2 (2022-09-29)
 
 ### Features Added